fix: change level progress to double

+++ src/main/java/ai/ready/ready/user/dto/ReadingStats.java
@@ -2,7 +2,7 @@ package ai.ready.ready.user.dto;
 
 public record ReadingStats (
     Integer level,
-    Integer levelProgress,
+    Double levelProgress,
     Integer numberOfReadBooks,
     Integer numberOfReadPages,
     Integer averageTimePerWeek
